requirements/server.txt
# ============================================
# DeepTutor Server — Full Dependencies (CLI + Web)
# ============================================
# Mirrors pyproject.toml `[project.optional-dependencies].server`.
# Keep in sync with pyproject.toml when adding/updating dependencies.
#
# Preferred install (from a source clone): pip install -e ".[server]"
# Use this file for Docker/CI when pyproject.toml/source code aren't yet available.

-r cli.txt

# --- API server ---
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# --- LLM provider runtime (unconditionally imported by provider_core) ---
loguru>=0.7.3,<1.0.0
json-repair>=0.57.0,<1.0.0

# --- Auth ---
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
aiosmtplib>=3.0.0
